Fix mocks
