+++ chainstate/storage/src/is_transaction_seal.rs
@@ -20,6 +20,6 @@ impl<'st, B: storage::Backend> Seal for crate::internal::StoreTxRo<'st, B> {}
 impl<'st, B: storage::Backend> Seal for crate::internal::StoreTxRw<'st, B> {}
 
 #[cfg(any(test, feature = "mock"))]
-impl Seal for crate::mock::MockStoreTxRo {}
+impl Seal for crate::mock_tests::MockStoreTxRo {}
 #[cfg(any(test, feature = "mock"))]
-impl Seal for crate::mock::MockStoreTxRw {}
+impl Seal for crate::mock_tests::MockStoreTxRw {}
