+++ chainstate/storage/src/lib.rs
@@ -18,7 +18,7 @@
 mod internal;
 mod is_transaction_seal;
 #[cfg(any(test, feature = "mock"))]
-pub mod mock;
+pub mod mock_tests;
 pub mod schema;
 
 use std::collections::BTreeMap;
